Resolve CARPLAST-33 "Diseño de mezcla"

See merge request microservices/cartonplast/cp-core-ws!27

+++ cp-core-vo/src/main/java/org/crsoft/cartonplast/vo/res/DieProductRes.java
@@ -24,6 +24,6 @@ public class DieProductRes implements Serializable {
     private Double area;
     private Double length;
     private Double width;
-    private Double gsmdis;
+    private Integer gsmdis;
     private CatalogStatusRes status;
 }
